filtering and sorting using query object

+++ Controllers/CommentController.cs
@@ -35,9 +35,11 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var comment = await _commentRepo.GetCommentByIdAsycn(id);
             if (comment == null)
             {
@@ -49,9 +51,11 @@ namespace api.Controllers
         }
 
         [HttpPost]
-        [Route("{stockId}")]
+        [Route("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentsDto createCommentsDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             bool stockExist = await _stockRepo.StockExist(stockId);
             if (!stockExist)
             {
@@ -65,11 +69,12 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto commentDto)
         {
 
-
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var commentModel = await _commentRepo.UpdateCommentAsync(id, commentDto.ToCommentFromUpdate());
             if (commentModel == null)
             {
@@ -81,15 +86,16 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var commentModel = await _commentRepo.DeleteCommentAsync(id);
-            if (commentModel == null){
-
+            if (commentModel == null)
+            {
                 return NotFound("Comment does not exist");
             }
-
             return Ok(commentModel.ToCommentDto());
         }
     }
